+++ sphinxcontrib/needs/roles/need_incoming.py
@@ -1,6 +1,8 @@
 from docutils import nodes
 from sphinx.util.nodes import make_refnode
 
+from sphinxcontrib.needs.utils import check_and_calc_base_url_rel_path
+
 try:
     from sphinx.errors import NoUri  # Sphinx 3.0
 except ImportError:
@@ -52,7 +54,9 @@ def process_need_incoming(app, doctree, fromdocname):
                         )
                     else:
                         new_node_ref = nodes.reference(target_need["id"], target_need["id"])
-                        new_node_ref["refuri"] = target_need["external_url"]
+                        new_node_ref["refuri"] = check_and_calc_base_url_rel_path(
+                            target_need["external_url"], fromdocname
+                        )
                         new_node_ref["classes"].append(target_need["external_css"])
 
                     node_link_container += new_node_ref